+++ Changelog.md
@@ -25,7 +25,7 @@
 - Added new skill tree for Jedi and new passive skills. ✔️
 - lmd_skillpoint_cost will have impact also on reset skill cost. Reseting skill is the same price as uprgrading. ✔️
 - Redesigned skill creation for mercenary. ✔️
-- Weapons select have more weapons now.
+- Weapons skill has more weapons available now. ✔️
 
 ## TODO